Moving request workflow to pages subdirectory
